105 create the resource api stack

+++ server/controller/apis/resourceController.js
@@ -18,9 +18,16 @@ const CompletedResource = require( '../../model/completedResource' );
 const Resource = require( '../../model/resource' );
 
 // set up file paths for user profile images
-let UPLOAD_PATH_BASE = path.resolve( __dirname, '..', '../../client' );
-let FRONT_END = process.env.FRONT_END_NAME;
-let RESOURCE_PATH = process.env.RESOURCE_IMAGE_PATH;
+const UPLOAD_PATH_BASE = path.resolve( __dirname, '..', '../../client' );
+const FRONT_END = process.env.FRONT_END_NAME;
+const RESOURCE_PATH = process.env.RESOURCE_IMAGE_PATH;
+
+// resource file path
+const resourceUploadPath = UPLOAD_PATH_BASE + "/" + FRONT_END + RESOURCE_PATH;
+
+// set the max image size for avatars and resource, topic and goal icons
+const maxSize = process.env.IMAGE_UPLOAD_MAX_SIZE;
+const maxSizeText = process.env.IMAGE_UPLOAD_MAX_SIZE_FRIENDLY_TEXT;
 
 
 /**
@@ -28,7 +35,56 @@ let RESOURCE_PATH = process.env.RESOURCE_IMAGE_PATH;
  * @param {*} req 
  * @param {*} res 
  */
-exports.getAllActiveResources = async function ( req, res ) {
+exports.getAllVisibleResources = async ( req, res ) => {
+    // get all the active resources by user
+    console.log("limit check:" + req.query.limit);
+
+    // get the auth user id from either the basic auth header or the session
+    let authUserId;
+    if( req.user ) {
+        authUserId = req.user.id;
+    }
+    else if( req.session.authUser ) {
+        authUserId = req.session.authUser.id;
+    }
+
+    console.log("auth user id; " + authUserId);
+    
+    if(authUserId > 0) {
+        let resources = await resourceService.getAllVisibleResources( authUserId, req.query.limit, req.query.offset );
+        res.set( "x-agora-message-title", "Success" );
+        res.set( "x-agora-message-detail", "Returned all resources" );
+        res.status( 200 ).json( resources );
+    }
+    else {
+        const message = ApiMessage.createApiMessage( 404, "Not Found", "Resource not found" );
+        res.set( "x-agora-message-title", "Not Found" );
+        res.set( "x-agora-message-detail", "Resource not found" );
+        res.status( 404 ).json( message );
+    }
+}
+
+
+/**
+ * 
+ * @param {*} req 
+ * @param {*} res 
+ */
+exports.getAllSharedResourcesForUser = async ( req, res ) => {
+    // get all the shared resources for this user
+    let sharedResources = await resourceService.getAllSharedResourcesForUser( req.session.authUser.id );
+
+    res.set( "x-agora-message-title", "Success" );
+    res.set( "x-agora-message-detail", "Returned all shared resources for user" );
+    res.status( 200 ).json( sharedResources );
+}
+
+/**
+ * 
+ * @param {*} req 
+ * @param {*} res 
+ */
+exports.getAllActiveResourcesForUser = async ( req, res ) => {
     // get all the active resources
     let resources = await resourceService.getAllActiveResourcesForOwner();
     
@@ -37,7 +93,7 @@ exports.getAllActiveResources = async function ( req, res ) {
     res.status( 200 ).json( resources );
 }
 
-exports.getResourceById = async function ( req, res ) {
+exports.getResourceById = async ( req, res ) => {
     // get all the active resources by user 
     let resource = await resourceService.getAllActiveResourcesForOwnerById( req.session.authUser.id, req.params.id );
     if(resource) {
@@ -55,7 +111,7 @@ exports.getResourceById = async function ( req, res ) {
     
 }
 
-exports.getAllResourcesForAuthUser = async function ( req, res ) {
+exports.getAllResourcesForAuthUser = async ( req, res ) => {
     // get all the resources for this owner
     let ownerResources = await resourceService.getAllResourcesForOwner( req.session.authUser.id, false );
 
@@ -109,18 +165,17 @@ exports.saveCompletedResource = async function( req, res ) {
     res.send();
 }
 
-
-
-exports.saveResourceImage = async function( req, res, resourceId, filename ) {
+exports.saveResourceImage = async( req, res, resourceId, filename ) => {
 
     // save image in db and delete old file  
     if( resourceId > 0 ) {
         resourceService.updateResourceImage( resourceId, filename ).then( ( rValue ) => {
 
-            if( rValue && rValue != 'resource-default.png' 
+            if( rValue && rValue.length > 0 && (rValue != 'resource-default.png' 
                 || rValue != 'notebook-pen.svg' 
                 || rValue != 'cell-molecule.svg' 
-                || rValue != 'code.svg' ) {
+                || rValue != 'code.svg' )) {
+                console.log("removing: " + UPLOAD_PATH_BASE + "/" + FRONT_END + RESOURCE_PATH + rValue)
                 fs.unlink( UPLOAD_PATH_BASE + "/" + FRONT_END + RESOURCE_PATH + rValue, ( err ) => {
                     if( err ) {
                         console.log( "[resourceController] file delete error status: " + err );
@@ -135,9 +190,7 @@ exports.saveResourceImage = async function( req, res, resourceId, filename ) {
     return true;
 }
 
-
-exports.saveResource = async function( req, res, redirect ) {
-
+exports.saveResource = async ( req, res, redirect ) => {
 
     let resource = Resource.emptyResource();
     resource.id = req.body.resourceId;
@@ -152,41 +205,172 @@ exports.saveResource = async function( req, res, redirect ) {
 
     // add changes from the body if they are passed
     resource.resourceType = req.body.resourceType;
-    resource.visibility = req.body.resourceVisibility;
+    resource.visibility = req.body.visibility;
     resource.resourceName = req.body.resourceName;
     resource.resourceDescription = req.body.resourceDescription;
+
     if(resource.resourceType == 3) {
+        
         resource.resourceContentHtml = req.body.embedded_submission_text_resource;
     }
     else {
-        resource.resourceContentHtml = req.body.resourceEditor;
+        // check to see if the incomping message format is from the UI or the API
+        if( req.body.resourceContentHtml ) {
+            resource.resourceContentHtml = req.body.resourceContentHtml;
+        }
+        else {
+            resource.resourceContentHtml = req.body.resourceEditor;
+        }
     }
     resource.resourceLink = req.body.resourceLink;
-    resource.active = ( req.body.resourceActive == "on" ) ? true : false;
-    resource.isRequired = ( req.body.isRequired == "on") ? true : false;
     
-    resource.ownedBy = req.session.authUser.id; 
-    resource = await resourceService.saveResource( resource );
+    // check to see if the incoming message format is from the UI form or the API
+    if( req.body.resourceActive ) {
+        resource.active = ( req.body.resourceActive == "on" ) ? true : false;
+    }
+    else if ( req.body.active ) {
+        resource.active = req.body.active;
+    }
+    
+    resource.isRequired = ( req.body.isRequired == "on" || req.body.isRequired == true ) ? true : false;
+    
 
-    if( resource ) {
-        req.session.messageType = "success";
-        req.session.messageTitle = "Resource Saved";
-        req.session.messageBody = "Resource " + resource.resourceName + " saved successfully!";
+    // get the user id either from the request user from basic auth in API call, or from the session for the UI
+    let authUserId;
+    if( req.user ) {
+        authUserId = req.user.id;
+    }
+    else if( req.session.authUser ) {
+        authUserId = req.session.authUser.id;
+    }
+    
+    if(authUserId > 0) {
+        resource.ownedBy = authUserId;
+
+        resource = await resourceService.saveResource( resource );
+
+        /**
+         * once the resource is saved, save the image if it is passed
+         */ 
+
+        // The UI needs to verify modifiction so that the image is not dropped if the user does not want to change it
+        if ( req.body.resourceModified && req.body.resourceModified != "false" && !req.files ) {
+            // do nothing we are going to keep the original file
+            console.log("resource trigger modification clause");
+        }
+        else if ( !req.files || Object.keys( req.files ).length === 0 ) {   // no files were uploaded       
+            // no files uploaded
+            if( resource.resourceType == 1 ) {
+                this.saveResourceImage( req, res, resource.id, 'notebook-pen.svg' );
+            }
+            else if ( resource.resourceType == 2 ) {
+                this.saveResourceImage( req, res, resource.id, 'cell-molecule.svg' );
+            }
+            else if( resource.resourceType == 3 ) {
+                this.saveResourceImage( req, res, resource.id, 'code.svg' );
+            }
+            else {
+                this.saveResourceImage( req, res, resource.id, 'resource-default.png' );
+            }
+        }
+        else {
+            // files included
+            const file = req.files.resourceImageField;
+            const timeStamp = Date.now();
+
+            // check the file size
+            if( file.size > maxSize ) {
+                console.log(`File ${file.name} size limit has been exceeded for resource`);
+
+                if(redirect) {
+                    req.session.messageType = "warn";
+                    req.session.messageTitle = "Image too large!";
+                    req.session.messageBody = "Image size was larger then " + maxSizeText + ", please use a smaller file. Your resource was saved without the image.";
+                    res.redirect( 303, '/dashboard' );
+                    return resource;
+                }
+                else {
+                    const message = ApiMessage.createApiMessage( 422, "Image too large", "Image size was larger then " + maxSizeText + ", please use a smaller file. Your resource was saved without the image.");
+                    res.set( "x-agora-message-title", "Image too large!" );
+                    res.set( "x-agora-message-detail", "Image size was larger then " + maxSizeText + ", please use a smaller file. Your resource was saved without the image.");
+                    res.status( 422 ).json( message );
+                }
+                
+            }
+            else if( resource ) {
+                await file.mv(resourceUploadPath + timeStamp + file.name, async (err) => {
+                    if (err) {
+                        console.log( "Error uploading profile picture : " + err );
+                        if(redirect) {
+                            req.session.messageType = "error";
+                            req.session.messageTitle = "Error saving image!";
+                            req.session.messageBody = "There was a error uploading your image for this resource. Your resource should be saved without the image.";
+                            res.redirect( 303, '/dashboard' );
+                            return resource;
+                        }
+                        else {
+                            const message = ApiMessage.createApiMessage( 422, "Error uploading image!", "There was a error uploading your image for this resource. Your resource should be saved without the image." );
+                            res.set( "x-agora-message-title", "Error saving image!" );
+                            res.set( "x-agora-message-detail", "There was a error uploading your image for this resource. Your resource should be saved without the image." );
+                            res.status( 422 ).json( message );
+                        }
+                    }
+                    else {
+                        await this.saveResourceImage( req, res, resource.id, timeStamp + file.name );
+                    }
+                });
+            }
+            else {
+                if(redirect) {
+                    req.session.messageType = "error";
+                    req.session.messageTitle = "Error saving image!";
+                    req.session.messageBody = "There was a error uploading your image for this resource. Your resource should be saved without the image.";
+                    res.redirect( 303, '/dashboard' );
+                    return resource;
+                }
+                else {
+                    const message = ApiMessage.createApiMessage( 422, "Error uploading image!", "There was a error uploading your image for this resource. Your resource should be saved without the image." );
+                    res.set( "x-agora-message-title", "Error saving image!" );
+                    res.set( "x-agora-message-detail", "There was a error uploading your image for this resource. Your resource should be saved without the image." );
+                    res.status( 422 ).json( message );
+                }
+            }
+        }
+
+        // redirect to the call the calling controller or return the resource if origin was an API call
+        if( resource ) {
+            if( redirect ) {
+                return resource;
+            }
+            else {
+                res.set( "x-agora-message-title", "Success" );
+                res.set( "x-agora-message-detail", "Returned all resources" );
+                res.status( 200 ).json( resource );
+            }
+        }
+        else {
+            if( redirect ) {
+                return resource;
+            }
+            else {
+                const message = ApiMessage.createApiMessage( 500, "Internal Server Error", "Error saving resource" );
+                res.set( "x-agora-message-title", "Internal Server Error" );
+                res.set( "x-agora-message-detail", "Error saving resource" );
+                res.status( 500 ).json( message );
+            }
+        }
     }
     else {
-        req.session.messageType = "error";
-        req.session.messageTitle = "Error saving Resource <br />";
-        req.session.messageBody = "There was a problem saving the resource. <br />";
-    }
-
-    if( redirect ) {
-        console.log( "resourceController.saveResource() - END - Redirect ");
-        return resource;
-    }
-    else {
-        console.log( "resourceController.saveResource() - END - Non-Redirect ");
-        res.setHeader( 'Content-Type', 'application/json' );
-        res.send(JSON.stringify(resource));
+        if( redirect ) {
+            return resource;
+        }
+        else {
+            const message = ApiMessage.createApiMessage( 401, "Not Authorized", "Not able to associate authorized user with the record" );
+            res.set( "x-agora-message-title", "Not Authorized" );
+            res.set( "x-agora-message-detail", "Not able to associate authorized user with the record" );
+            res.status( 401 ).json( message );
+        }
+        
     }
     
 }
